refactor: relocate engine core files
