+++ engine/builders/coreBuilder.ts
@@ -1,6 +1,6 @@
-import { GameEngine, gameEngineDependencies, gameEngineToken } from '@engine/gameEngine'
-import { TurnScheduler, turnSchedulerDependencies, turnSchedulerToken } from '@engine/turnScheduler'
-import { EngineInitializer, engineInitializerDependencies, engineInitializerToken } from '@engine/engineInitializer'
+import { GameEngine, gameEngineDependencies, gameEngineToken } from '@core/gameEngine'
+import { TurnScheduler, turnSchedulerDependencies, turnSchedulerToken } from '@core/turnScheduler'
+import { EngineInitializer, engineInitializerDependencies, engineInitializerToken } from '@core/engineInitializer'
 import { Container } from '@ioc/container'
 import type { Container as IContainer } from '@ioc/types'
 import { MessageBus, messageBusDependencies, messageBusToken } from '@utils/messageBus'
