+++ engine/main.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { ContainerBuilder, IContainerBuilder } from '@builders/containerBuilder'
-import { gameEngineToken, IGameEngine } from '@engine/gameEngine'
+import { gameEngineToken, IGameEngine } from '@core/gameEngine'
 import { Container } from '@ioc/container'
-import { turnSchedulerToken } from '@engine/turnScheduler'
+import { turnSchedulerToken } from '@core/turnScheduler'
 import './styling/reset.css'
 import './styling/variables.css'
 import './styling/engine.css'
