+++ tests/engine/containerBuilder.test.ts
@@ -1,12 +1,12 @@
 import { describe, it, expect, vi } from 'vitest'
 import { ContainerBuilder } from '@builders/containerBuilder'
-import { gameEngineToken } from '@engine/gameEngine'
+import { gameEngineToken } from '@core/gameEngine'
 import { messageBusToken } from '@utils/messageBus'
 import { messageQueueToken } from '@utils/messageQueue'
-import { GameEngine } from '@engine/gameEngine'
+import { GameEngine } from '@core/gameEngine'
 import { MessageBus } from '@utils/messageBus'
 import { MessageQueue } from '@utils/messageQueue'
-import { TurnScheduler, turnSchedulerToken } from '@engine/turnScheduler'
+import { TurnScheduler, turnSchedulerToken } from '@core/turnScheduler'
 import { ServiceProvider, serviceProviderToken } from '@providers/serviceProvider'
 import { dataPathProviderToken } from '@providers/configProviders'
 import { GameDataProvider, gameDataProviderToken } from '@providers/gameDataProvider'
