+++ tests/engine/coreBuilder.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, vi } from 'vitest'
 import { CoreBuilder } from '@builders/coreBuilder'
-import { engineInitializerToken } from '@engine/engineInitializer'
-import { gameEngineToken } from '@engine/gameEngine'
-import { turnSchedulerToken } from '@engine/turnScheduler'
+import { engineInitializerToken } from '@core/engineInitializer'
+import { gameEngineToken } from '@core/gameEngine'
+import { turnSchedulerToken } from '@core/turnScheduler'
 import { keyboardEventListenerToken } from '@utils/keyboardEventListener'
 import { messageBusToken } from '@utils/messageBus'
 import { messageQueueToken } from '@utils/messageQueue'
