+++ tests/engine/engineInitializer.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest'
-import { EngineInitializer } from '../../engine/engine/engineInitializer'
+import { EngineInitializer } from '../../engine/core/engineInitializer'
 import type { ILogger } from '@utils/logger'
 import { START_GAME_ENGINE_MESSAGE, SWITCH_PAGE } from '../../engine/messages/system'
 import { postMessageActionToken } from '../../engine/actions/postMessageAction'
@@ -14,7 +14,6 @@ import type { IActionManager } from '../../engine/managers/actionManager'
 import type { IMapManager } from '../../engine/managers/mapManager'
 import type { IVirtualKeyProvider } from '../../engine/providers/virtualKeyProvider'
 import type { IVirtualInputProvider } from '../../engine/providers/virtualInputProvider'
-import type { ITurnManager } from '../../engine/managers/turnManager'
 import type { Game } from '../../engine/loader/data/game'
 import type { ITurnManager } from '../../engine/managers/turnManager'
